combined files into path_builder
