--- global_paths/mock_global_path.csv
@@ -1,2 +1,0 @@
-latitude,longitude
-54.20058314396191,-165.82740730800987
