+++ global_paths/path.csv
@@ -1,6 +1,3 @@
 latitude,longitude
-48.406961,-126.958008
-48.78479,-129.089355
 48.158391,-130.253906
-47.820163,-132.1875
 48.121368,-137.022955
