--- global_paths/sample_interpolated_path.csv
@@ -1,19 +1,0 @@
-latitude,longitude
-47.16727951986584,122.16404796040233
-47.3343177531939,122.32912846558237
-47.50111183916594,122.49525262682697
-47.66765887690069,122.66243167767188
-47.83395592481907,122.83067697497849
-48.0,123.0
-48.16727710590574,123.1639545490304
-48.33431395292454,123.32897811023895
-48.50110763508846,123.49508245831032
-48.66765520451942,123.66227950271488
-48.833953670744414,123.83058128898794
-49.0,124.0
-49.167273943178515,124.16385741727373
-49.33430895433564,124.32882172779209
-49.50110208264115,124.49490542435441
-49.66765033341839,124.6621211485995
-49.83395066740474,124.83048169252501
-50.0,125.0
